Add auth dto class and rename user dto folder to more excplicity
