+++ src/main/java/com/dunice/Vitushkin_Advanced_REST_Server/dto/user/RegisterUserDto.java
@@ -1,4 +1,4 @@
-package com.dunice.Vitushkin_Advanced_REST_Server.dto.userDto;
+package com.dunice.Vitushkin_Advanced_REST_Server.dto.user;
 
 import com.dunice.Vitushkin_Advanced_REST_Server.exception.ErrorsMsg;
 import jakarta.validation.constraints.Email;
