+++ src/main/java/com/dunice/Vitushkin_Advanced_REST_Server/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package com.dunice.Vitushkin_Advanced_REST_Server.mapper;
 
-import com.dunice.Vitushkin_Advanced_REST_Server.dto.userDto.LoginUserDto;
-import com.dunice.Vitushkin_Advanced_REST_Server.dto.userDto.RegisterUserDto;
+import com.dunice.Vitushkin_Advanced_REST_Server.dto.user.LoginUserDto;
+import com.dunice.Vitushkin_Advanced_REST_Server.dto.user.RegisterUserDto;
 import com.dunice.Vitushkin_Advanced_REST_Server.models.User;
 import org.mapstruct.Mapper;
